refactor type and method name

+++ newrelic-agent/src/main/java/com/newrelic/agent/config/SpanEventsConfig.java
@@ -56,8 +56,8 @@ public class SpanEventsConfig extends BaseConfig {
         return maxSamplesStored;
     }
 
-    public void setMaxSamplesStoredByServerProp(Long harvestLimit) {
-        this.maxSamplesStored = harvestLimit.intValue();
+    public void setMaxSamplesStoredByServerProp(int harvestLimit) {
+        this.maxSamplesStored = harvestLimit;
     }
 
     public int getTargetSamplesStored() {
